undo type change to maintain compatability with pbc 1.7.x based versions

+++ src/main/kotlin/io/provenance/client/grpc/BaseReq.kt
@@ -29,7 +29,7 @@ data class BaseReq(
     val signers: List<BaseReqSigner>,
     val body: TxBody,
     val chainId: String,
-    val gasAdjustment: Float? = null,
+    val gasAdjustment: Double? = null,
     val feeGranter: String? = null
 ) {
